create test module for concentric_circle
